Добавит сброс ошибок валидации при закрытии формы. Вынесет код валидации в отдельный модуль. Вынесет модули, соответствующие одной зоне ответстенности в отдельные папки. Вынесет функции приведения к числу с плавающей запятой и округлению числа в util
