+++ js/upload-form/scale.js
